+amount update

+++ src/main/java/com/lyaev/accounts/model/OperationJSON.java
@@ -2,9 +2,11 @@ package com.lyaev.accounts.model;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OperationJSON {
     private String accountName;
-    private Double summ;
+    private BigDecimal summ;
     private byte isDebit;
 }
